Refactored, can move king, passes all tests
Moved a lot of repeated movement code into the Piece class itself, this supports Rook, Bishop, and Queen. Added a movement flag in piece which is used in Pawn and soon King. The king can now move as expected - I decided not to code so defensively this time and the function worked as expected. All tests still pass after refactoring.